[WIP] kvmtool: don't rely on $HOME
kvm__set_dir() called in main() and kvm__get_dir() rely on $HOME. But in
some environments (such as starting lkvm through systemd-run), $HOME is
undefined. This causes bind() to use a socket path containing "(null)"
and this fails. The current code does not check errors returned by
realpath().

Symptoms:

| bind: No such file or directory
|   Error: Failed adding socket to epoll
|  Warning: Failed init: kvm_ipc__init
|
|  Fatal: Initialisation failed

This patch was first reported on https://github.com/coreos/rkt/pull/1394

It is marked as work-in-progress because ideally /root should not be
used to store socket files.